pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "wids-wildfire-threat"
version = "0.1.0"
description = "WiDS 2026 — Predicting Time-to-Threat for Evacuation Zones Using Survival Analysis"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pandas>=2.1.0",
    "polars>=0.20.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.2.0",
    "lifelines>=0.28.0",
    "scikit-survival>=0.23.0",
    "mlflow>=2.10.0",
    "structlog>=24.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
